refactor: addToCart 액션 분리
