+++ src/basic/features/cart/actions/addToCart.ts
@@ -1,0 +1,49 @@
+import { findProductById } from '../../product/actions';
+import { Product } from '../../product/types/Product';
+
+const addToCart = (
+  CartItemsView: HTMLDivElement,
+  selectedItemId: Product['id'],
+  productList: Product[],
+  callback: () => void,
+) => {
+  const itemToAdd = findProductById(selectedItemId, productList);
+  if (itemToAdd && itemToAdd.quantity > 0) {
+    const item = document.getElementById(itemToAdd.id);
+    if (item) {
+      const newQty =
+        parseInt(item.querySelector('span').textContent.split('x ')[1]) + 1;
+      if (newQty <= itemToAdd.quantity) {
+        item.querySelector('span').textContent =
+          itemToAdd.name + ' - ' + itemToAdd.price + '원 x ' + newQty;
+        itemToAdd.quantity--;
+      } else {
+        alert('재고가 부족합니다.');
+      }
+    } else {
+      const newItem = document.createElement('div');
+      newItem.id = itemToAdd.id;
+      newItem.className = 'flex justify-between items-center mb-2';
+      newItem.innerHTML =
+        '<span>' +
+        itemToAdd.name +
+        ' - ' +
+        itemToAdd.price +
+        '원 x 1</span><div>' +
+        '<button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="' +
+        itemToAdd.id +
+        '" data-change="-1">-</button>' +
+        '<button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="' +
+        itemToAdd.id +
+        '" data-change="1">+</button>' +
+        '<button class="remove-item bg-red-500 text-white px-2 py-1 rounded" data-product-id="' +
+        itemToAdd.id +
+        '">삭제</button></div>';
+      CartItemsView.appendChild(newItem);
+      itemToAdd.price--;
+    }
+    callback();
+  }
+};
+
+export { addToCart };
