+++ src/basic/features/cart/actions/index.ts
@@ -1,0 +1,3 @@
+import { addToCart } from './addToCart';
+
+export { addToCart };
