+++ src/basic/main.basic.ts
@@ -1,3 +1,4 @@
+import { addToCart } from './features/cart/actions/addToCart';
 import { productList } from './features/product';
 import {
   findProductById,
@@ -145,45 +146,11 @@ function updateStockInfo() {
 }
 main();
 AddToCartButton.addEventListener('click', function () {
-  var selectedItemId = SelectView.value;
-  const itemToAdd = findProductById(selectedItemId, productList);
-  if (itemToAdd && itemToAdd.quantity > 0) {
-    var item = document.getElementById(itemToAdd.id);
-    if (item) {
-      var newQty =
-        parseInt(item.querySelector('span').textContent.split('x ')[1]) + 1;
-      if (newQty <= itemToAdd.quantity) {
-        item.querySelector('span').textContent =
-          itemToAdd.name + ' - ' + itemToAdd.price + '원 x ' + newQty;
-        itemToAdd.quantity--;
-      } else {
-        alert('재고가 부족합니다.');
-      }
-    } else {
-      var newItem = document.createElement('div');
-      newItem.id = itemToAdd.id;
-      newItem.className = 'flex justify-between items-center mb-2';
-      newItem.innerHTML =
-        '<span>' +
-        itemToAdd.name +
-        ' - ' +
-        itemToAdd.price +
-        '원 x 1</span><div>' +
-        '<button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="' +
-        itemToAdd.id +
-        '" data-change="-1">-</button>' +
-        '<button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="' +
-        itemToAdd.id +
-        '" data-change="1">+</button>' +
-        '<button class="remove-item bg-red-500 text-white px-2 py-1 rounded" data-product-id="' +
-        itemToAdd.id +
-        '">삭제</button></div>';
-      CartItemsView.appendChild(newItem);
-      itemToAdd.price--;
-    }
+  const selectedItemId = SelectView.value;
+  addToCart(CartItemsView, selectedItemId, productList, () => {
     calculateCartItems();
     lastSelectedItemValue = selectedItemId;
-  }
+  });
 });
 CartItemsView.addEventListener('click', function (event) {
   var targetElement = event.target;
